feat: update dev-tools documentation
