+++ packages/@o3r/artifactory-tools/src/cli/pr-artifact-cleaner.ts
@@ -2,20 +2,19 @@
 
 import { program } from 'commander';
 import { Headers, Options } from 'request';
-import * as request from 'request-promise-native';
 import * as winston from 'winston';
 
 program
   .description('Clean pr artifacts from artifactory repositories')
   .requiredOption('-u, --artifactory-url <artifactoryUrl>', 'Artifactory URL')
-  .option<number>('-d, --duration-kept <numberOfDays>', 'Only artifacts which are older and have not been downloaded during this duration (in days) will be deleted.', (v) => +v, 1)
-  .option<number>('-pr, --pr-versions <prVersions>', 'Number of pr versions that will be kept.', (v) => +v, 1)
-  .option('-r, --repository <repositoryName>', 'Artifact repository to clean up.', 'dga-maven-built-adt-nce')
-  .option('-p, --path <path>', 'Artifact paths to cleanup use matcher from AQL language. Be careful that the path do not include release artifacts.', 'com/amadeus/retailing/*-PR-*')
-  .option('--dry-run', 'List all files that should be deleted without actually deleting them')
+  .requiredOption('-r, --repository <repositoryName>', 'Artifact repository to clean up.')
+  .requiredOption('-p, --path <path>', 'Artifact paths to clean up (using matcher from AQL language). Be careful not to include release artifacts in the path.')
+  .option<number>('-d, --duration-kept <numberOfDays>', 'All artifacts older than this value (in days) will be deleted.', (v) => +v, 1)
+  .option<number>('-n, --pr-builds <prBuilds>', 'Number of PR build artifacts that will be kept.', (v) => +v, 1)
+  .option('--dry-run', 'List all files that would be deleted without actually deleting them', false)
   .option('-a, --api-key <apiKey>', 'Artifactory API Key of the user that can be used to log in')
-  .option('-b, --basicAuth <base64>', 'Base 64 encoding of username:password (password already encrypted from artifactory UI)')
-  .option('-v, --verbose', 'To display the executed AQL query')
+  .option('-b, --basicAuth <base64>', 'Base64 encoding of username:password (password already encrypted from artifactory UI)')
+  .option('-v, --verbose', 'Display the executed AQL query')
   .parse(process.argv);
 
 const programOptions = program.opts();
@@ -44,7 +43,7 @@ const logger = winston.createLogger({
 });
 
 if (!programOptions.basicAuth && !programOptions.apiKey) {
-  logger.error('Authentication is mandatory, please specify a base 64 encoded user:password with b parameter or an ApiKey with -a parameter');
+  logger.error('Authentication is mandatory, please specify a base64 encoded user:password with the -b parameter or an ApiKey with the -a parameter');
   process.exit(1);
 }
 if (programOptions.basicAuth && programOptions.apiKey) {
@@ -56,7 +55,7 @@ const authHeader: Headers = programOptions.basicAuth ? { Authorization: `Basic $
 let url: string = programOptions.artifactoryUrl;
 url += (url.endsWith('/') ? '' : '/') + 'api/search/aql';
 const ageInDays: number = programOptions.durationKept;
-const prVersions: number = programOptions.prVersions;
+const prBuilds: number = programOptions.prBuilds;
 const repository: string = programOptions.repository;
 const path: string = programOptions.path;
 const options: Options = {
@@ -79,59 +78,27 @@ const options: Options = {
 logger.debug(`AQL search executed : ${options.body}`);
 logger.info(`Url called : ${url}`);
 
-/**
- * @param result
- * @param result.repo
- * @param result.path
- * @param result.name
- * @param pathWithoutBuildNumber
- */
-function getUniqueId(result: { repo: string; path: string; name: string }, pathWithoutBuildNumber: string) {
-  let mapId: string;
-  if (result.path.indexOf('.') !== -1 && result.path.indexOf('-') !== -1) {
-    const splittedPath = result.path.split('/');
-    const partialPathWithBuildNumber = splittedPath.find((currentPath) => currentPath.indexOf('.') !== -1);
-    if (partialPathWithBuildNumber) {
-      const splittedPartialPathWithDot = partialPathWithBuildNumber.split('.');
-      splittedPartialPathWithDot.splice(splittedPartialPathWithDot.length - 1, 1);
-      mapId = splittedPartialPathWithDot.join('.');
-    } else {
-      const splittedName = result.name.split('.');
-      const splittedNameDash = splittedName[splittedName.length - 2].split('-');
-      splittedNameDash.splice(0, 1);
-      splittedName[splittedName.length - 2] = splittedNameDash.join('-');
-      const nameWithoutBuildNumber = splittedName.join('.');
-      mapId = pathWithoutBuildNumber + '/' + nameWithoutBuildNumber;
-    }
-  } else {
-    mapId = pathWithoutBuildNumber + result.name;
-  }
-  return mapId;
-}
-
 // eslint-disable-next-line @typescript-eslint/no-floating-promises
 (async () => {
   logger.info(`Requesting old artifacts using  ${url}`);
   let responseSearch: any;
+  let responseSearchObj: { results: { repo: string; path: string; name: string }[] };
   try {
-    responseSearch = await request.post(url, options).promise();
-  } catch {
-    logger.warn('No result found %o', responseSearch);
+    responseSearch = await fetch(url, {method: 'POST', headers: authHeader, body: options.body});
+    responseSearchObj = await responseSearch.json();
+  } catch (e) {
+    logger.warn('No result found %o', e);
     process.exit(0);
   }
-  const responseSearchObj: { results: { repo: string; path: string; name: string }[] } = JSON.parse(responseSearch);
   /** uris will contain the list of all artifacts that need to be deleted */
   const mapOfKeptItems = new Map<string, number[]>();
   const mapOfKeptResult = new Map<string, { repo: string; path: string; name: string }[]>();
   const resultToDelete: { repo: string; path: string; name: string }[] = [];
   const sortedResult = responseSearchObj.results.sort((a, b) => b.name.localeCompare(a.name));
   for (const result of sortedResult) {
-    const splittedPathBySlash = result.path.split('/');
-    const folderwithBuildNumber = splittedPathBySlash.filter((value) => value.indexOf('.') !== -1)[0];
-    const splittedPath = folderwithBuildNumber.split('.');
-    const pathWithoutBuildNumber = splittedPath.slice(0, -1).join('.');
-    const currentBuildNumber = +splittedPath[splittedPath.length - 1];
-    const mapId: string = getUniqueId(result, pathWithoutBuildNumber);
+    const splitPath = result.name.split('.');
+    const mapId = splitPath.slice(0, -2).join('.');
+    const currentBuildNumber = +splitPath[splitPath.length - 2];
     const buildNumbers = mapOfKeptItems.get(mapId);
     if (!buildNumbers) {
       mapOfKeptItems.set(mapId, [currentBuildNumber]);
@@ -144,10 +111,10 @@ function getUniqueId(result: { repo: string; path: string; name: string }, pathW
         isBuildNumberHigherThanExisting = isBuildNumberHigherThanExisting && (value < currentBuildNumber);
       });
       if (!isBuildNumberAlreadyInMap) {
-        if (buildNumbers.length >= prVersions && !isBuildNumberHigherThanExisting) {
+        if (buildNumbers.length >= prBuilds && !isBuildNumberHigherThanExisting) {
           resultToDelete.push(result);
         } else {
-          if (buildNumbers.length >= prVersions && isBuildNumberHigherThanExisting) {
+          if (buildNumbers.length >= prBuilds && isBuildNumberHigherThanExisting) {
             const buildNumberToRemove = buildNumbers.shift();
             if (buildNumberToRemove) {
               const resultsToRemove = mapOfKeptResult.get(`${mapId}${buildNumberToRemove}`);
@@ -172,11 +139,13 @@ function getUniqueId(result: { repo: string; path: string; name: string }, pathW
   }
   logger.debug('Map of build that will be kept: %o', mapOfKeptItems);
 
-  const filesToDelete = resultToDelete.map((data) => (programOptions.artifactoryUrl as string) + repository + '/' + data.path + '/' + data.name);
+  const filesToDelete = resultToDelete.map(
+    (data) => (programOptions.artifactoryUrl as string) + (programOptions.artifactoryUrl.endsWith('/') ? '' : '/') + repository + '/' + data.path + '/' + data.name
+  );
   for (const uri of filesToDelete) {
     logger.info(`Deleting ${uri}...`);
     if (!programOptions.dryRun) {
-      const response = await request.delete(uri, options).promise();
+      const response = await fetch(uri, {method: 'DELETE', headers: authHeader});
       logger.info(response);
     }
   }
