--- packages/@o3r/artifactory-tools/src/helpers/gavc-response/index.ts
@@ -1,1 +1,0 @@
-export * from './gavc-response.interfaces';
