--- packages/@o3r/artifactory-tools/src/public_api.ts
@@ -1,1 +1,0 @@
-export * from './helpers/index';
