+++ packages/@o3r/dev-tools/src/cli/comment-pr.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 import { spawnSync } from 'node:child_process';
 
-console.warn('This CLI is deprecated, please use `o3r-comment-pr` from the package `@o3r/artifactory-tools`');
+console.warn('This CLI is deprecated, please use `o3r-comment-pr` from the package `@o3r/azure-tools`');
 
 spawnSync('o3r-comment-pr', process.argv.slice(2), {
   shell: true,
